Daily target check

+++ diet-service/src/main/kotlin/com/rmp/diet/dto/water/WaterLogUploadDto.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class WaterLogUploadDto(
-    val volume: Float,
+    val volume: Double,
 ): SerializableClass
